+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Absolute: +1.03MB
     - Relative: +25.6%
 - Add missing tests for anagrafica NISECI csv parsing
+- Add license info button
 
 ### Changed
 
@@ -25,7 +26,7 @@
 - Fixed wrong message in error reports for `check_riferimento_niseci_reader`
 - Refactor theme loading into `core::controller::GuiTheme::load_and_set()`
 - Use double default font height on `GuiTheme::Light`
-  - At startup (`MainState` prep)
+  - At startup (`MainState`, `ConsoleView` prep)
   - On theme switch
 
 ## [0.0.11] - 2025-03-12